Split node device APIs to libvirtApi/nodeDev.js
